Ensure event is finished before transaction
When `Appsignal::Transaction.complete_current!` is called within a
`Appsignal.instrument` block the event is not properly closed. This is
because the event metadata is only set at the end of the block, after
the transaction is "completed" and closed for modifications.

See:
https://github.com/appsignal/appsignal-ruby/blob/172bf1f37f9ae260a5759e8a31daa7729896a35a/lib/appsignal/helpers/instrumentation.rb#L421-L424

Most likely the event will arrive in AppSignal as an "unknown" event.

By moving the `Appsignal::Transaction.complete_current!` call outside
the `Appsignal.instrument` block the event can be properly finished
before the transaction is "completed" and closed for modifications.